Remove role name that could be misleading
Drupal allows for significant flexibility in roles and permissions.
It doesn't provide a 'Content Manager' role right out of the box,
though, and site administrators make a variety of decisions about
who has access to what.

So at the risk of sounding pedantic, this edit changes `Content
Manager` to something more general, in hopes of sparing avoidable
confusion.